fetch list of configs to change using type hint checking, cached once
