+++ compute_space/src/compute_space/core/default_apps.py
@@ -25,13 +25,13 @@ from typing import Any
 import attr
 
 from compute_space.config import Config
-from compute_space.core.apps import all_manifest_permissions_v2
 from compute_space.core.apps import clone_and_read_manifest
 from compute_space.core.apps import insert_and_deploy
 from compute_space.core.apps import move_clone_to_app_temp_dir
 from compute_space.core.apps import validate_manifest
 from compute_space.core.logging import logger
 from compute_space.core.manifest import AppManifest
+from compute_space.core.manifest import all_manifest_permissions_v2
 from compute_space.core.manifest import parse_manifest
 
 MAX_RETRY_ATTEMPTS = 3
