+++ compute_space/src/compute_space/core/installer.py
@@ -16,12 +16,12 @@ import sqlite3
 import attr
 
 from compute_space.config import Config
-from compute_space.core.apps import all_manifest_permissions_v2
 from compute_space.core.apps import clone_with_github_fallback
 from compute_space.core.apps import insert_and_deploy
 from compute_space.core.apps import move_clone_to_app_temp_dir
 from compute_space.core.apps import validate_manifest
 from compute_space.core.auth.permissions_v2 import Grant
+from compute_space.core.manifest import all_manifest_permissions_v2
 
 # Service URL the v2 service proxy matches against to dispatch to this
 # module instead of resolving a normal provider app.
