+++ compute_space/src/compute_space/core/manifest.py
@@ -1,7 +1,13 @@
+import functools
+import json
 import os
 import re
 import tomllib
+from typing import Annotated
 from typing import Any
+from typing import get_args
+from typing import get_origin
+from typing import get_type_hints
 
 import attr
 import cattrs
@@ -9,6 +15,7 @@ from packaging.specifiers import InvalidSpecifier
 from packaging.specifiers import SpecifierSet
 
 from compute_space.core.auth.permissions_v2 import Grant
+from compute_space.core.auth.permissions_v2 import PermissionRecord
 from compute_space.core.logging import logger
 
 _SHORTNAME_RE = re.compile(r"^[a-z][a-z0-9_-]{0,31}$")
@@ -123,26 +130,35 @@ class ServiceConsumes:
 
 
 @attr.s(auto_attribs=True, frozen=True)
+class SettingLabel:
+    """Marks an :class:`AppManifest` field as shown in the update review diff,
+    under ``group``, displayed as ``text``. See :func:`manifest_setting_labels`."""
+
+    group: str
+    text: str
+
+
+@attr.s(auto_attribs=True, frozen=True)
 class AppManifest:
     # [app]
     name: str
-    version: str
-    description: str = ""
-    authors: list[str] = attr.Factory(list)
+    version: Annotated[str, SettingLabel("App", "Version")]
+    description: Annotated[str, SettingLabel("App", "Description")] = ""
+    authors: Annotated[list[str], SettingLabel("App", "Authors")] = attr.Factory(list)
 
     # [runtime]
-    runtime_type: str = "serverfull"
+    runtime_type: Annotated[str, SettingLabel("App", "Runtime type")] = "serverfull"
 
     # [runtime.container]
-    container_image: str = ""
-    container_port: int = 0
-    container_command: str | None = None
-    port_mappings: list[PortMapping] = attr.Factory(list)
-    capabilities: list[str] = attr.Factory(list)
-    devices: list[str] = attr.Factory(list)
+    container_image: Annotated[str, SettingLabel("Container", "Image")] = ""
+    container_port: Annotated[int, SettingLabel("Container", "Container port")] = 0
+    container_command: Annotated[str | None, SettingLabel("Container", "Command")] = None
+    port_mappings: Annotated[list[PortMapping], SettingLabel("Ports", "Port mappings")] = attr.Factory(list)
+    capabilities: Annotated[list[str], SettingLabel("Container", "Linux capabilities")] = attr.Factory(list)
+    devices: Annotated[list[str], SettingLabel("Container", "Devices")] = attr.Factory(list)
     # `--shm-size` (in MiB).  0 = use podman's default (64 MiB).
     # Apps doing serious browser work (jibri) need ~2 GiB minimum.
-    shm_mb: int = 0
+    shm_mb: Annotated[int, SettingLabel("Container", "Shared memory (MB)")] = 0
     # Use the host's network namespace instead of pasta.  Required for apps
     # that do IP forwarding (VPN servers like WireGuard).  Pasta proxies
     # individual TCP/UDP connections but cannot forward routed packets from
@@ -152,42 +168,64 @@ class AppManifest:
     # connect to any port on the host, including other apps' loopback-bound
     # ports, the router, and local services.  Only use for apps that
     # genuinely need raw network access (VPNs, transparent proxies).
-    network_host: bool = False
+    network_host: Annotated[bool, SettingLabel("Container", "Host networking")] = False
 
     # [routing]
-    health_check: str | None = None
-    public_paths: list[str] = attr.Factory(list)
+    health_check: Annotated[str | None, SettingLabel("Routing", "Health check")] = None
+    public_paths: Annotated[list[str], SettingLabel("Routing", "Public paths")] = attr.Factory(list)
 
     # [[links]]
-    links: list[AppLink] = attr.Factory(list)
+    links: Annotated[list[AppLink], SettingLabel("Routing", "Links")] = attr.Factory(list)
 
     # [resources]
-    memory_mb: int = 128
-    cpu_cores: float = 0.1
-    gpu: bool = False
+    memory_mb: Annotated[int, SettingLabel("Resources", "Memory (MB)")] = 128
+    cpu_cores: Annotated[float, SettingLabel("Resources", "CPU cores")] = 0.1
+    gpu: Annotated[bool, SettingLabel("Resources", "GPU")] = False
 
     # [data]
-    sqlite_dbs: list[str] = attr.Factory(list)
-    app_data: bool = True
-    app_temp_data: bool = False
-    app_archive: bool = False
-    access_vm_data: bool = False
+    sqlite_dbs: Annotated[list[str], SettingLabel("Data", "SQLite databases")] = attr.Factory(list)
+    app_data: Annotated[bool, SettingLabel("Data", "Permanent data")] = True
+    app_temp_data: Annotated[bool, SettingLabel("Data", "Temporary data")] = False
+    app_archive: Annotated[bool, SettingLabel("Data", "Archive data")] = False
+    access_vm_data: Annotated[bool, SettingLabel("Data", "Access VM data")] = False
     # Granular cross-app data flags (preferred):
-    access_all_app_data: bool = False
-    access_all_archive: bool = False
+    access_all_app_data: Annotated[bool, SettingLabel("Data", "Access all app data")] = False
+    access_all_archive: Annotated[bool, SettingLabel("Data", "Access all archive")] = False
     # Convenience shorthand: equivalent to access_all_app_data + access_all_archive.
-    access_all_data: bool = False
+    access_all_data: Annotated[bool, SettingLabel("Data", "Access all data")] = False
 
     # [services.v2]
-    provides_services_v2: list[ServiceProvides] = attr.Factory(list)
+    provides_services_v2: Annotated[list[ServiceProvides], SettingLabel("Services", "Services provided")] = (
+        attr.Factory(list)
+    )
 
-    # [[services.v2.consumes]]
+    # [[services.v2.consumes]] — diffed separately by manifest_newly_declared_permissions_v2,
+    # not part of the review label set.
     consumes_services_v2: list[ServiceConsumes] = attr.Factory(list)
 
     # [app] metadata
-    hidden: bool = False
+    hidden: Annotated[bool, SettingLabel("App", "Hidden")] = False
 
     raw_toml: str = ""
+
+
+@functools.cache
+def manifest_setting_labels() -> dict[str, SettingLabel]:
+    """``{field_name: SettingLabel}`` for every :class:`AppManifest` field annotated
+    with :class:`SettingLabel`, in declaration order. Drives the update review diff
+    (see :func:`manifest_settings_changes`) so it stays in sync with the manifest
+    schema without a separately maintained field list. Cached: pure function of a
+    fixed class, computed once regardless of how many manifests are diffed."""
+    hints = get_type_hints(AppManifest, include_extras=True)
+    labels: dict[str, SettingLabel] = {}
+    for field in attr.fields(AppManifest):
+        hint = hints[field.name]
+        if get_origin(hint) is not Annotated:
+            continue
+        for extra in get_args(hint)[1:]:
+            if isinstance(extra, SettingLabel):
+                labels[field.name] = extra
+    return labels
 
 
 def _validate_devices(devices: list[Any]) -> list[str]:
@@ -454,3 +492,141 @@ def parse_manifest(repo_path: str) -> AppManifest:
         raw_bytes = f.read()
 
     return parse_manifest_from_string(raw_bytes.decode("utf-8"))
+
+
+@attr.s(auto_attribs=True, frozen=True)
+class PermissionGrant:
+    """A single permission grant: which service and what payload."""
+
+    service_url: str
+    grant: Grant
+
+
+def all_manifest_permissions_v2(manifest: AppManifest) -> list[PermissionGrant]:
+    """Build a permissions_v2_grants list that approves every permission declared in the manifest."""
+    grants: list[PermissionGrant] = []
+    for perm in manifest.consumes_services_v2:
+        for grant_payload in perm.grants:
+            grants.append(PermissionGrant(service_url=perm.service, grant=grant_payload))
+    return grants
+
+
+def _permission_key(service_url: str, grant_payload: Grant) -> tuple[str, str]:
+    """Normalized identity for a (service, grant) pair.
+
+    Uses the same ``json.dumps(..., sort_keys=True)`` serialization that
+    :func:`grant_permission_v2` stores, so a declared grant and a stored grant
+    compare equal regardless of dict key order.
+    """
+    return (service_url, json.dumps(grant_payload, sort_keys=True))
+
+
+def manifest_ungranted_permissions_v2(
+    manifest: AppManifest,
+    granted: list[PermissionRecord],
+) -> list[PermissionGrant]:
+    """The permissions a manifest declares that are NOT already granted.
+
+    Single source of truth for the "which manifest permissions still need owner
+    approval" diff, shared by the app-detail page (post-install display) and the
+    update/reload gate (so an app update can't silently pick up newly declared
+    permissions). ``granted`` is the app's current ``permissions_v2`` rows
+    (e.g. from :func:`get_all_permissions_v2`).
+
+    Duplicate manifest declarations collapse to a single entry, and each new
+    permission is returned only once even if declared multiple times.
+    """
+    granted_keys = {_permission_key(rec.service_url, rec.grant) for rec in granted}
+    ungranted: list[PermissionGrant] = []
+    seen: set[tuple[str, str]] = set(granted_keys)
+    for perm in manifest.consumes_services_v2:
+        for grant_payload in perm.grants:
+            key = _permission_key(perm.service, grant_payload)
+            if key in seen:
+                continue
+            seen.add(key)
+            ungranted.append(PermissionGrant(service_url=perm.service, grant=grant_payload))
+    return ungranted
+
+
+def manifest_newly_declared_permissions_v2(
+    manifest: AppManifest,
+    granted: list[PermissionRecord],
+    previous_manifest_raw: str | None,
+) -> list[PermissionGrant]:
+    """The permissions an update *newly* declares: in the new manifest, not already
+    granted, and not declared by the previously-deployed manifest.
+
+    This is the update gate's delta. Gating on it — rather than on grant state
+    (see :func:`manifest_ungranted_permissions_v2`) — means a permission the owner
+    deliberately revoked is NOT re-surfaced for approval on an unrelated update:
+    it was declared by the previous manifest too, so it isn't part of the delta.
+    Only grants the app *author* actually added are gated.
+
+    Scope is permissions only; other manifest changes (ports, resources, image,
+    …) always apply on update regardless of this diff.
+
+    ``previous_manifest_raw`` is the currently-deployed manifest text (``apps.manifest_raw``).
+    When it is absent or unparseable, this falls back to the full grant-state diff
+    (equivalent to :func:`manifest_ungranted_permissions_v2`) — the safe default,
+    since it never silently grants something the owner hasn't seen.
+    """
+    ungranted = manifest_ungranted_permissions_v2(manifest, granted)
+    if not previous_manifest_raw:
+        return ungranted
+    try:
+        previous_manifest = parse_manifest_from_string(previous_manifest_raw)
+    except ValueError:
+        return ungranted
+    previously_declared = {
+        _permission_key(pg.service_url, pg.grant) for pg in all_manifest_permissions_v2(previous_manifest)
+    }
+    return [pg for pg in ungranted if _permission_key(pg.service_url, pg.grant) not in previously_declared]
+
+
+@attr.s(auto_attribs=True, frozen=True)
+class SettingChange:
+    """A single manifest setting whose value changed between two deployments."""
+
+    group: str
+    label: str
+    old: str
+    new: str
+
+
+def _render_setting_value(value: Any) -> str:
+    """One-line human rendering of a manifest field value for the update diff."""
+    if isinstance(value, bool):
+        return "yes" if value else "no"
+    if isinstance(value, list):
+        return ", ".join(_render_setting_value(v) for v in value) if value else "(none)"
+    if attr.has(type(value)):
+        return json.dumps(attr.asdict(value), sort_keys=True)
+    return str(value)
+
+
+def manifest_settings_changes(manifest: AppManifest, previous_manifest_raw: str | None) -> list[SettingChange]:
+    """Grouped diff of :class:`SettingLabel`-annotated manifest settings changed vs
+    the previously-deployed manifest. Empty when there's no parseable previous
+    manifest; permissions are excluded (diffed separately, see
+    :func:`manifest_newly_declared_permissions_v2`)."""
+    if not previous_manifest_raw:
+        return []
+    try:
+        previous = parse_manifest_from_string(previous_manifest_raw)
+    except ValueError:
+        return []
+    changes: list[SettingChange] = []
+    for field, setting_label in manifest_setting_labels().items():
+        old_val = getattr(previous, field)
+        new_val = getattr(manifest, field)
+        if old_val != new_val:
+            changes.append(
+                SettingChange(
+                    group=setting_label.group,
+                    label=setting_label.text,
+                    old=_render_setting_value(old_val),
+                    new=_render_setting_value(new_val),
+                )
+            )
+    return changes
